email template

+++ database/migrations/2017_08_07_002940_create_sell_coins_table.php
@@ -20,7 +20,7 @@ class CreateSellCoinsTable extends Migration
             $table->string('buyer_email')->nullable();
             $table->bigInteger('buyer_phone')->nullable();
             $table->bigInteger('amount_dollar')->nullable();
-            $table->bigInteger('amount_btc')->nullable();
+            $table->float('amount_btc')->nullable();
             $table->integer('rate')->nullable();
             $table->timestamps();
         });
